rename files and remove old logger / old firebase config
