+++ backend/src/modules/notification/routes/preferences.router.ts
@@ -1,5 +1,5 @@
 import { Router } from "express";
-import { preferencesController } from "../controllers/preferencesController";
+import { preferencesController } from "../controllers/preferences.controller";
 const preferencesRoutes = Router();
 
 preferencesRoutes.get(`/:userId`, preferencesController.getUserPreferences);
