+++ backend/src/modules/search/routes/search.routes.ts
@@ -1,6 +1,6 @@
 import { Router } from "express";
 
-import { SearchController } from "../controllers/searchController";
+import { SearchController } from "../controllers/search.controller";
 const searchRoutes = Router();
 
 searchRoutes.get("/users/", SearchController.searchUsers);
