+++ backend/src/modules/subscription/routes/subscription.routes.ts
@@ -1,5 +1,5 @@
 import express from 'express';
-import subscriptionController from '../controllers/subscriptionController';
+import subscriptionController from '../controllers/subscription.controller';
 import { authenticateToken } from '../../../shared/middleware/auth';
 
 const router = express.Router();
